requirements.txt
# XHS Content Studio - Python 依赖
# 安装: pip install -r requirements.txt

# 爬虫依赖
playwright>=1.40.0
selenium>=4.15.0
openpyxl>=3.1.0
requests>=2.31.0

# 数据库同步
supabase>=2.0.0
python-dotenv>=1.0.0

# 后端 API 服务
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
